Improve typing for math libraries. Move and rename custom-math-js.test.ts.
