+++ src/diagram/custom-mathjs.ts
@@ -3,9 +3,19 @@ import { create, simplifyDependencies, unitDependencies,
   parseDependencies, evaluateDependencies, toDependencies,
   createUnitDependencies } from "mathjs";
 
-import { customUnitsArray, deleteUnits } from "./custom-mathjs-units";
+import { customUnitsArray, deleteUnits, IUnit } from "./custom-mathjs-units";
 
-function createMath() {
+export interface IMathLib {
+  simplify: math.Simplify;
+  unit: any;
+  parse: math.ParseFunction;
+  createUnit: (name: string, definition?: string | math.UnitDefinition | undefined, options?: math.CreateUnitOptions | undefined) => math.Unit;
+  evaluate: (expr: math.MathExpression | math.MathExpression[], scope?: object | undefined) => any;
+  number: (value?: string | number | boolean | math.MathArray | math.Matrix | math.Unit | math.BigNumber | math.Fraction | null | undefined) => number | math.MathArray | math.Matrix;
+  isUnit: (x: unknown) => x is math.Unit;
+  Unit: any;
+}
+function createMath(): IMathLib {
   // This reduces the size of the bundle, see:
   // https://mathjs.org/docs/custom_bundling.html
   const m = create({ simplifyDependencies, unitDependencies, parseDependencies,
@@ -22,7 +32,7 @@ function createMath() {
   const mathUnit = (m as any).Unit;
 
   deleteUnits.forEach((u: string) => mathUnit.deleteUnit(u));
-  customUnitsArray.forEach((u: any) => m.createUnit(u.unit, u.options));
+  customUnitsArray.forEach((u: IUnit) => m.createUnit(u.unit, u.options));
 
   const isAlphaOriginal = mathUnit.isValidAlpha;
   mathUnit.isValidAlpha = function (c: string): boolean {
@@ -37,7 +47,8 @@ function createMath() {
     evaluate: m.evaluate,
     number: m.number,  
     isUnit: m.isUnit,
-    Unit: mathUnit };
+    Unit: mathUnit
+  };
 }
 
 export { createMath, };
